Big change -
Created full CRUD reminders feature that allows a date to be set. If the date is past or "today" then a notification is displayed on the dashboard. Otherwise, a smiley face is displayed.

Also updated homepage design and content to make more sense with the new features added to this app.