Cleaned up stats section using string splitting for formatting in HTML.
Styled date due on reminders a bit.  Added IF statements for styling depending on if a reminder is due or late.

Added stats to nav bar.

Moved javascript for progress page to that page only, as it was loading throughout the entire site and causing errors due to unknown elements.

Added transparent screwfix logo and restyled homepage for it.